% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game}
\alias{game}
\title{Game}
\format{A data frame with ... rows and 4 variables:
\describe{
  \item{game_date}{Date of game}
  \item{crowd}{Announced Attendance}
  \item{referee_name}{Initial and Last name}
  \item{game_id}{Unique Game ID}
}}
\usage{
game
}
\description{
Basic information including, date, attendance and referee
}
\keyword{datasets}
