% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game_team}
\alias{game_team}
\title{Games by Team}
\format{A data frame with 20252 rows and 4 variables:
\describe{
  \item{team_game_id}{Unique Team Game ID}
  \item{game_id}{Unique Game ID}
  \item{venue}{Stadium. A - Away; H - Home}
  \item{team}{Team name}
}}
\usage{
game_team
}
\description{
Dataset of Team games including venue
}
\keyword{datasets}
