% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{games}
\alias{games}
\title{Games}
\format{A data frame with 10127 rows and 4 variables:
\describe{
  \item{game_date}{Date of game}
  \item{crowd}{Announced Attendance}
  \item{referee_name}{Initial and Last name}
  \item{game_id}{Unique Game ID}
}}
\usage{
games
}
\description{
Basic information including, date, attendance and referee
}
\keyword{datasets}
