% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goals}
\alias{goals}
\title{Goals}
\format{A data frame with ... rows and 6 variables:
\describe{
  \item{player_game_id}{Unique Player Game ID of Scorer}
  \item{time}{time in game. Added time at end of each half is included in 45 and 90 minute figure }
  \item{method}{Method. R - Right foot; L -  Left Foot; H - Head(or other body part)}
  \item{place}{Location on Pitch. G - 6 yard box; P - Others in Penalty Area; O - Outside Penalty area}
  \item{play}{Play. C - Corner; D - Direct Free Kick; I - Indirect FK; O - Open Play; P - Penalty; T - Throwin}
}}
\usage{
goals
}
\description{
Dataset of Goals including information on time scored, method
place and play
}
\keyword{datasets}
