% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manager_team}
\alias{manager_team}
\title{Manager by Team}
\format{A data frame with 367 rows and 5 variables:
\describe{
  \item{manager_joined}{Date of Hiring}
  \item{manager_left}{Date of Leaving}
  \item{manager_team_id}{Unique ID for Manager's stint on Team}
  \item{team}{Team name}
  \item{manager_id}{Unique Manager ID}
}}
\usage{
manager_team
}
\description{
Time period of manager's stint with team
}
\keyword{datasets}
