% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{managers}
\alias{managers}
\title{Managers}
\format{A data frame with 224 rows and 2 variables:
\describe{
  \item{manager_id}{Unique Manager ID}
  \item{manager_name}{Manager's name}
}}
\usage{
managers
}
\description{
Dataset of Managers
}
\keyword{datasets}
