% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_game}
\alias{player_game}
\title{Game by Player}
\format{A data frame with 330287 rows and 9 variables:
\describe{
  \item{player_game_id}{Unique Player Game ID}
  \item{team_game_id}{Team Game ID}
  \item{player_team_id}{Player Team ID}
  \item{time_off}{Time of game when substituted. 0 otherwise}
  \item{time_on}{Time of game entering as sub. ) otherwise}
  \item{caution}{Y - Yellow; P - 2 Yellows; R - Straight Red; Z Yellow then Red;
  otherwise NULL. N.B. Few odd values to resolve }
  \item{own_goals}{Scored on Own Team}
  \item{missed_penalties}{Missed a penalty. Follow-up goals not included}
  \item{start}{Starter in Game indicated by value > 0}
}}
\usage{
player_game
}
\description{
Individual Game details for player
}
\keyword{datasets}
