% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{player_scoring_method}
\alias{player_scoring_method}
\title{total goals a player has scored with right foot, left-foot or header}
\usage{
player_scoring_method(id)
}
\arguments{
\item{id}{player_id}
}
\value{
a dataframe of goals by scoring method
}
\description{
total goals a player has scored with right foot, left-foot or header
}
\examples{
player_scoring_method("KANEH")
player_scoring_method("SHEAREA")
}
