% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_team}
\alias{player_team}
\title{Player by Team}
\format{A data frame with ... rows and 8 variables:
\describe{
  \item{player_id}{Unique Player ID}
  \item{player_joined}{Date first signed - estimated for academy players}
  \item{player_left}{Date player was transferred or became free agent}
  \item{permanent}{1 - Permanent, 0 - On-loan}
  \item{fee_in}{Fee paid on transfer in  in 1000 pounds. Could be best estimate. 99 - unknown}
  \item{fee_out}{Fee paid on transfer out in 1000 pounds. Could be best estimate. 99 - unknown}
  \item{player_team_id}{Unique ID for Player's stint on Team}
  \item{team}{Team name}

  }}
\usage{
player_team
}
\description{
Player details by team they played for including dates with
team and transfer fees
}
\keyword{datasets}
