% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{players}
\alias{players}
\title{Premier League players 1992-2018}
\format{A data frame with 4690 rows and 7 variables:
\describe{
  \item{player_id}{Unique Player ID}
  \item{first_name}{First name - may be blank}
  \item{last_name}{Last name - may be nickname e.g Juninho}
  \item{birth_date}{Birth Date}
  \item{birth_city}{Town or City of birth, if known}
  \item{birth_country}{Country of Birth - not nationality}
  \item{position}{Usual Playing Position G-Goalkeeper, D-Defender, M-Midfielder, F-Forward}
}}
\usage{
players
}
\description{
Basic data for all players who have appeared in
the Premier League. May include some on rosters
who never played
}
\keyword{datasets}
