% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_functions.R
\name{team_max_game_score}
\alias{team_max_game_score}
\title{most goals a team has scored in game}
\usage{
team_max_game_score(id)
}
\arguments{
\item{id}{team}
}
\value{
an integer of maximum goals scored in game
}
\description{
most goals a team has scored in game
}
\examples{
team_max_game_score("Man. Utd.")
team_max_game_score("Barnsley")
}
