% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assists}
\alias{assists}
\title{Assists}
\format{A data frame with ... rows and 2 variables:
\describe{
  \item{player_game_id}{Unique Player Game ID of Assister}
  \item{goal_id}{Goal which Assist was provided for}

}}
\usage{
assists
}
\description{
Dataset of Assists for Goals (maximum 2 per goal)
}
\keyword{datasets}
